fix(prod-rag): cast numpy float scores to native float before API response
ChromaDB distances and FlashRank rerank scores are numpy.float32, which
pydantic/fastapi cannot JSON-serialize, causing 500s on /inference
(PydanticSerializationError: Unable to serialize unknown type:
<class 'numpy.float32'>). Cast to float() at the point each score is
computed in chromadb_client.py and reranker.py.

+++ prod-rag/src/chromadb_client.py
@@ -57,7 +57,7 @@ class ChromaDBClient:
                 "id": chunk_id,
                 "text": text,
                 "metadata": metadata or {},
-                "score": 1 / (1 + distance),
+                "score": float(1 / (1 + distance)),
                 "source": "chroma",
             }
             for chunk_id, text, metadata, distance in zip(
